requirements.txt
packaging 
wheel
fire
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
arrow==1.3.0
asttokens==2.4.0
async-timeout==4.0.3
attributedict==0.3.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.1.0
blessings==1.7
blinker==1.6.2
blis==0.7.11
boto3==1.28.59
botocore==1.31.59
bravado==11.0.3
bravado-core==6.1.0
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
cmake==3.27.6
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
colour-runner==0.1.1
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
coverage==7.3.2
cramjam==2.7.0
cycler==0.12.0
cymem==2.0.8
DataProperty==1.0.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.6.0
deepspeed==0.10.3
dill==0.3.7
distlib==0.3.7
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
evaluate==0.4.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fastparquet==2023.8.0
filelock==3.12.4
flash-attn==2.3.2
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
inspecta==0.1.3
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
lit==17.0.2
llvmlite==0.41.0
lm-eval==0.3.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbformat==5.9.2
neptune==1.7.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numba==0.58.0
numexpr==2.8.7
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==0.28.1
optimum==1.13.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathvalidate==3.2.0
pathy==0.10.2
patsy==0.5.3
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pybind11==2.11.1
pycountry==22.3.5
pydantic==1.10.13
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
pyodbc==4.0.39
pyparsing==3.1.1
pyproject-api==1.6.1
PySocks==1.7.1
pytablewriter==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.6.0
rootpath==0.1.1
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.10.3
s3transfer==0.7.0
sacrebleu==1.5.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
selenium==4.14.0
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.27.2
svgwrite==1.4.3
swagger-spec-validator==3.0.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
texttable==1.7.0
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tox==4.11.3
tqdm==4.66.1
tqdm-multiprocess==0.0.11
traitlets==5.11.2
transformers==4.34.0
trio==0.22.2
trio-websocket==0.11.1
triton==2.1.0
trl==0.7.1
typepy==1.3.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
urllib3==1.26.17
validators==0.22.0
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.8
weasel==0.3.2
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.3
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.21.0
optimum
auto-gptq